docs: [#17] model 파일 분리
