+++ Phew/Phew/Home/DailyRoutineRecord.swift
@@ -30,47 +30,3 @@ class DailyRoutineRecord {
         "\(date.monthAndDay())-\(dailyRoutineType.rawValue)"
     }
 }
-
-@Model
-class DailyRoutineResponse {
-    @Attribute(.unique) var id: UUID
-    var dailyRoutineResponseType: DailyRoutineResponseType
-    var answerText: String?
-    var answerScore: Int?
-    
-    enum DailyRoutineResponseType: String, Codable  {
-        case text
-        case score
-        case none
-    }
-    
-    init(id: UUID, dailyRoutineResponseType: DailyRoutineResponseType, answerText: String? = nil, answerScore: Int? = nil) {
-        self.id = id
-        self.dailyRoutineResponseType = dailyRoutineResponseType
-        self.answerText = answerText
-        self.answerScore = answerScore
-    }
-}
-
-@Model
-class DailyRoutineTask {
-    @Attribute(.unique) var id: UUID
-    var taskType: DailyRoutineTaskType
-    var title: String
-    var subTitle: String?
-    var imageName: String?
-    
-    enum DailyRoutineTaskType: String, Codable {
-        case question
-        case slider
-        case quote
-    }
-    
-    init(id: UUID, taskType: DailyRoutineTaskType, title: String, subTitle: String? = nil, imageName: String? = nil) {
-        self.id = id
-        self.taskType = taskType
-        self.title = title
-        self.subTitle = subTitle
-        self.imageName = imageName
-    }
-}
